Add stop place insert when creating stop on map
Earlier when we created a stop, we only created the scheduled_stop_point to
routes and lines database, but we need to also add an stop place entry to tiamat.
After that we need to update the newly created scheduled_stop_point's stop_place_ref
to point to the newly created stopPlace.
NOTE: These three queries are needed, but they are not in a transaction yet. So if one of the latter
queries fail, there is no rollbacks. This could've been all in one "mutation" instead of three separate, but
I think that could give the impression that they are done in a transaction which it isn't since they are actually
in different databases and GQL does not support transactions between multiple databases.
Also a minor thing but moved GQL_INSERT_STOP to useCreateStop.ts.