Change cypress intercept config to speed up the tests
When running cypress on local machine it takes 10-15s per test
(.visit() command) to load up before the test actually starts. This
seems to be mostly because of cypress was intercepting every single
call made. There was a comment about "we should only intercept the
hasura calls, but it doesnt work". But at least now it seems to work
just fine. And also this sped up the tests majorly which feels amazing.