Refactor e2e database clear logic
We had some antipatterns in our database clearing logic. Each test cleared the database
resources which they were about to use in their tests before the tests and after the tests.
But because there is no guarantee that the resource clear is done after the tests, this caused problems because
it would mean that some dependant resource was not cleared and another test would conflict on data clear. More info
about this on the ticket https://dev.azure.com/hslfi/JORE%204.0/_workitems/edit/40300.
And when the tests did run, we lost one benefit of cypress which is that cypress would in the exact state after the test
is ran so the developer can continue checking things out which is obviously helpful.

So what was done
- all after() and afterEach() blocks were removed from tests
- Each test now resets the databases in beforeEach block (except some login tests etc.)
- In the database reset we truncate all (currently) necessary routes and lines db tables, we run the timetables
truncate sql file, and for tiamat we query all the stop place ids and use the tiamat API to delete all stop places. These
probably need to change in the future when we start to use more data for tiamat, but for now this is sufficient.
- Removed the now unused db-utils.ts file
- Removed the now unused removeFromDbHelper method
- Removed some extra 'emptyDownloadsFolder' tasks from some tests that did not download anything.
- Added /downloads to .gitignore, since now we are not clearing that up after tests