Use dataset in timetablesImport and rework
- To use the dataset in timetablesImport we had to create new import files for the dataset.
The new import files of course now has timetables for both directions, but also 2 vehicles
instead of one. The import file data now covers a bit more functionalities.
- Fix some minor typo in testId (dateType -> dayType)
- Aded a temporary `data-testidtemp´ to VehicleJourneyRow, which will be replaced after all
of the old usage of the other testid is removed, but did not want to bloat this PR anymore.
So the old testid used diretion and label as row identifier, but there are multiple rows with those identifiers,
so it is not a working solution.
- Used a bit more spesific naming with the import files
- Replaced all unnecessary cy.visit() -commands with UI navigation inside the tests
- Added new pageObject to folder `timetables/import`, this could be used for all the import pageObjects later
- Added a `route_validity_start` to timetables dataset, because timetables-data-inserter defaults this to
2023-06-01 and we want it to start from January (also some tests already replace these timetables and they can't
replace the existing timetables if they start later or at the same time as the imported timetable)
- End note: The import files might still have some data fields that are not realistic or should be modified, but
they also might not have any effect on jore4 whatsoever.