Remove temporary test id from VehicleJourneyRow
After we removed the `getVehicleJourneyRow(routeLabel: string, routeDirection: string)` usage from e2e tests,
we can remove it from the component test and remove it completely and replace it with the data-testidtemp, which
doesn't have any identifying parameters. For context: the route label and direction was not enough to identify one row anyway,
because there could be many rows with the same label and direction. Also we have been refactoring these usages towards
a guideline that we get all rows from a list and then pick the nth row. With this approach we also can be sure that the lists are
in correct order