Use dataset in createRoute and refactor
- Replaced dynamic testIds with static ones (dynamic testIds are a bit problematic and e.g. using dynamic testIds in lists
ignore the order of the list completely)
- Added pageObject RouteStopsOverlayRow
- Renamed `addToJourneyPatternButton` testId to `toggleInJourneyPatternButton` since the button has two different functionalities depending
on whether the stop belongs to the journey pattern or not.
- Removed mapModal pageObject's `createRoute` method completely. It supported very narrow usage and we should not hide that much UI test logic
behind very complex and long methods. Also removed `createStopNextToAnotherStop` method as unused and also if in the future we merge the stopMarkers
as the map's `feature`, we can't use the mapMarkers in the tests at all.
- Removed removeStopsFromRoute method from RouteStopsOverlay for many reasons. We should have this functionality in the tests, not in method. Also this
was named `removeStopsFromRoute`, even though it would just as well add a stop to a route if the chosen stop was not used on the route.
- Removed some unused methods from RouteEditor, some of these wouldn't even work anymore since the map update (snapping line & edit handles
are not working the same way) and also
- End note: As iterative approach I did not change/fix everything at once, so there are still things to enhance here, probably some methods might be unused, but
small parts at a time.