Add dropdown menu for adding stop areas and terminals
The SimpleRoundButton does have some problems still and often (always?)
requires some workarounds to get it positioned correctly.
In this case we use setting width for the container (larger than actual button)
and justify-content properties on parents to center it properly.
Looks alright in this case, but might require similar hacks if used elsewhere.
The root cause seems to be aspect-ratio, which does set the width correctly
based on height, but unfortunately this does _not_ expand the parent height
which then messes up alignment.