Refactor Line routes from table to list and rename
- From accessibility point of view things that are not
tables, should not be created by table elements. Also tables add unnecessary complexity to
the structure and readability compared to a simple list and list items.
- Naming was a bit off in the old structure: we had RouteStopsTable (now LineRouteList) then a single row of that table
was RouteStopsSection (now LineRouteListItem). RouteStopsSection had ExpandableRouteRow (now just RouteRow)
component and the RouteStopsRow (now RouteStopListItem).
- The element structure was also semantically off, since the RouteStopsSection returned a <tbody> and not a single <tr>
and we actually populated the same table with two different data: ExpandableRouteRow and RouteStopsRow.
- The list item rows are now structured with CSS grid, which might not be the absolute best here, but has some nice benefits
with smaller screens. But this could also be achieved using percentage widths.
- The white lines and gray lines between list items (border) is now created with a pseudo element to achieve the designed visuals for the
border. This can be changed if we implement the left "colour coded" alerts with some other way. By using different coloured border for
top and left will create a indirect line to the corner of the element and at least I couldn't work my way around it.
- All the tests have been matched with the new names and new testIds, so nothing more has been done there.