Prefer Tiamat GraphQL API for fetching intricate stop area details
Only use the direct DB connection for fetching the absolute minimum we need.
Using it causes all kinds of problems, eg. with typing.
Also it is generally better the less we tie ourselves to the DB implementation.

Now, we only fetch the stop area id + location, which we need to show it on map,
and remaining details are fetched only when needed = area is clicked.
Stops on map have been implemented in a similar way.
Plan is to use this same data when editing the stop area (TODO, soon).