Add loader and disable stop actions dropdown while loading
DISCLAIMER: This is not a complete solution to an issue where user could
overwrite one stop action with another. So there is a possibility that
user removes a stop from route and then removes another stop from route
before the route stops list has updated from the previous removal. In this
case the second request will overwrite the first one.
The reason for this is that the list does not get updated instantly after
the update request is done (not sure if apollo is just very slow here).
We do also fetch the route data after the update, but even if we waited for
that request to be done, it wouldn't be enough. The route fetch request
gets response before the route stop list gets updated.

Anyways this loading indicator is still an UX enhancement.