Use e2e base dataset in routeStops
This commit also has some pageObject rework, since there were a bit of methods on 'wrong'
pageObjects.
- Some test logic was moved to tests from the page objects. Some of it can be added back to
page object if wanted, but the methods need to be on the correct page objects. But for now,
they were just removed as iterative change.
- Added `container::` to RouteStopListItem's testId so that we can also get all list items
with `cy.get('[data-testid^="RouteStopListItem::container"]')`. It would probably be best if
we refactored the dynamic testIds as well, but that would've bloated this PR, so this solution
works again as an iterative move.
- One test ('Checking "Use Hastus place" ...') was completely removed, since it's content is
tested within the next test.