Remove dynamic testId from RouteStopListItem
Now the toggleRouteSection and dropdownAction also follows the 'within' pattern.
In the changes made to hastusExport there is now a bit awkward flow in the
'should show an error, when neither the last stop nor the first stop is a timing point' test
where we can't set both of the timingSettings in the same "getNthLineRouteListItem(0)" due to
the component structure. So we have the `<TimingSettingsModal /> actually in LineDetailsPage.tsx
so it can't be accessed within the same LineRouteListItem. This is probably something that should be refactored
in a way that the modal is not rendered so far from the actual component it is part of (RouteStopListItem/StopActionsDropdown).