e2e: Improve Map zoom in action
Instead of focusing the map, which could result in a click on the map,
which could select something and alter the maps state, and then firing
'+' -keyboard inputs. We now just click the on screen zoom-in button.

Also all map panning and zoom events are debounced on out code. So an
extra 1s wait is added to all single zoom-in clicks.