e2e: Improve Map zoom in action
Instead of focusing the map, which could result in a click on the map,
which could select something and alter the maps state, and then firing
'+' -keyboard inputs. We now just click the on screen zoom-in button.

Additionally wait 1 second between zoom-operations to let the map
settle in the new state, as else it might ignore the next zoom action.