Fix unauthorized map login loop
The login loop issue is actually a two fold:
We have protected routes from unauthorized users by redirecting them to front page for login,
in case the userInfo can't be retrieved. But we also have a "temporary quality of life" addition so
that if a GQL request fails, we reload the page, which is added so that when developers are stopping
and starting the dependencies, they lose the authorisation from the auth service, but the userInfo
is still there, so the pages are accessible, but the GQL's return nothing, and with the reload it actually
fetches the userInfo again -> not authorized -> redirect to front page.

But since mapModal is not an individual page and is in fact a modal and visible anywhere if the queryString
'mapOpen=true' is present, and if the map is open on top of the root page so localhost:3300?mapOpen=true
we were not redirecting the user to front page, because front page does not require auth. But the temp QoL
causes the page to reload -> So in the end the page reloads -> sends GQLs -> gets unauthed -> reloads..

The second part is that even if we removed the "reload QoL addition" in client.ts which would fix the loop,
it would still mean that the map can be opened on top of front page, it would just have an infinite loader spinning
and no data. So probably best to "protect" it as well for now, before we get some better solutions.

- This change broke some e2e tests, because we had some random network waits.
  - Removed those waits since they are not good practice anways.
- Adjusted the click points on tests to be more far away from the loading indicator.
  - The original click now clicked on top of the loading indicator which broke the tests. Optimally the
  loading indicator could be waited here, but there is still some inconsitencies with the loading indicators and
  not part of this PR.