Update map-matching pinned image and fix post request url
Map-matching service was recently updated to spring boot 3, which
changed the default behaviour of trailing slashes. So this `/bus/`
request stopped working since we don't have the trailing slash explicitly
defined in the API. So in this commit we remove it. Here is the source for the
spring boot change https://github.com/spring-projects/spring-boot/wiki/Spring-Boot-3.0-Migration-Guide#spring-mvc-and-webflux-url-matching-changes