Move MapLoader to MapModal component
The map is not it's own page and is in fact a modal/dialog. So if you "click outside of the dialog" it will close,
and because the map loader was not inside the map modal component, it is considered as "outside of dialog" and
if you click the loader, the map will close.

Don't know or can't remember why map loader was outside of map modal component in the first place, but it probably
should be inside anyway.