Fix url pathname from disappearing when opening the map
useUrlQuery's setMultipleParametersToUrlQuery's pathname was defaulted
to a an empty string when it should've been undefined. Under the hood it uses
useNavigate hook and an empty value is a valid value for the navigate function
and it will empty the pathname.
So we change the default value to undefined, which results in no changes to
pathname.