Only allow editing of a single Stop Area page block at a time
Prevent editing of multiple data blocks simultaneously. Backend does
not support partial mutations. If member stops field is left empty it
deletes all member stops from the area. Thus on the background,
editing the details section also sends in the current list of known
member stops as of the time editing started. Editing member stops
performs similarly, sending in the known values of the details
sections at the time editing started.

Thus we want to limit the user to editing a single block of data at a
time.