E2E: Allow expecting a Success Toast but fail on other types
Instead of simply waiting for a Success Toast, allow one to wait for
any Toast, and then assert that it is of a correct type.

Simply waiting for Success Toast can result in unhelpful screenshots
if the waited operation fails and shows an Error Toast instead of a
Success. But due to Cypress waiting for the Success to appear the
Error Toast will eventually dissappear from the page, and won't be
included in the screen shots.