Add `areEqual` utility - dispatches to obj's own equals methods
Lodash's `isEqual` which performs a simple deep comparison of objects
based on their own enumerable properties. But some times classes
provide their own semantics on equality. For example Luxons DateTime
does so and in some cases `a.equals(b)` !== `isEqual(a, b)`.

This helper checks the objects being compared for custom
implementations of `equals` functions and dispatched the comparison to
that function instead.