Add memoize function implementation with a single item cache
Lodash does provide a memoization function, but that has an infinite
cache size, thus by default it causes a memory leak.

Inspired by https://github.com/alexreardon/memoize-one, but a
simplified version of it, with a very basic functionality. Could be
replaced by memoize-one in future, if a more robust implementation is
needed.

By default uses Object.is to compare the function parameters with, but
allows for alternative comparators too.