Fix openOnMap -hooks to not mutate URL on call
`useOpenStopOnMap` and `useShowStopAreaOnMap` -hooks rely on
`useObservationDateQueryParam` -hook to provide the currently active
observationDate for `openMapWithParameters` function. By default
`useObservationDateQueryParam` performs a page navigation to set the
observationDate to the default value if it not already present in the
current page's search string.

`useOpenStopOnMap` is also present on the StopSearch StopTableRow
component, which results in the URL being manipulated in loop, each
loop rewriting the previous iterations query param with new version,
once for each row present in a initial page render of an stop search
result table.

Fixed the hooks, to only read in the current observationDate, never to
set it during render.