E2E: Fix 'Should show arrival times' test
A recent change to LoadingWrapper seems to have broken the test.
Some sort of an timing change in the render cycle has taken place and
previously the test checked that there were 0 times on the screen, and
that test passed because at the time there was a spinner on the screen
instead, but now the actual contents that would have been displayed
after the spinner get to the screen before cypress has a chance to
verify the non existence of the arrival times.

Changed the test to assume that the data does actually get onto the
screen.