Add utilities needed to map Copied InfoSpots to new shelters
Shelters don't have specific ID that could be used to match the old
version to the new one. Thus we need to equality checking in order to
attach the InfoSpots to them once they are copied.

By default the matching happens based on the Shelters properties, but
if there are multiple identical Shelters a index based 🤞🏻 solution is
used instead. But technically the Shelters are not ordered in anyway
on the DB levels, thus the indexes might chance. Tough in practice
PostgreSQL should return data in INSERT order, and the insertion the
Shelter copies should also happen in specific order.