Preemptively add StopRegistryNamedEntitity form bits.
These were still needed in the initial iteration of for the new Stop
creation modal, but were later dropped from the design. But these will
be needed when we update the StopArea cration modal in the future. So
as to not lose this already written valid code to /dev/null, the files
are here now.