E2E: Improve map/createRoute tests
Previously the form was filled in pieces: base details and then the
priority and validity period were filled in manually afterwards. But
the code in charge of filling in the base details also tries to fill
in the priority and validity periods. This split resulted in the
endDate/indefinite filds to flicker between states, sometimes
resulting in test failures.

Now the form is filled-in in one swell swoop.