Improve Stop copying
Coder's previous assumption was that one could alter the Quays
included in a StopPlace by simply altering the contents of quays field
in the mutation payload. But that is not case. Instead on Tiamat's end
the quays field is always populated from the DB, and then any new
quays are added from the mutation payload or existing ones changes, if
there is a QuayInput fragment with a matching ID in the payload. Thus,
omitting a Quay from the update payload, does not result in it getting
deleted from StopPlace.

Instead to delete a Quay from a StopPlace a special `deleteQuay`
mutation must be called.