E2E: Add component to get coordinates for the map
Adds a component that can visualize the mouse events on the page. Can
be used to get proper click coordinates for use in map related tests,
or to visualize what cypress is doing while running the tests.

The component needs to be enabled manually on the code when needed, by
setting `enableCypressCoordinateHelper` to `true` in file:
`ui/src/pages/index.tsx`