Make basic date[time] formatting locale independent
Instead of relying on the global I18Next instance to always inject the
current locale into the DateTime object before formatting, a single
fixed format is used for all languages.

The formatting already had a fixed format for the date portion, but
the time portion was dependent on the active locale. The time portion
has now been fixed to the same format that was already used for when
Finnish as the active locale 'H.mm'. This format also aligns the
recent decision to prefer UK formats in English language texts, and UK
people are aware of the 24H clock system.

Additional helper function have been provided to allow for proper
locale aware DateTime formatting in the future, but for that the
locale must be properly retrieved and provided from the calling scope.