Replace `parseI18nField` with reactive alternative
Old `parseI18nField` got the locale from a globa reference to the
I18Next object instance, and thus was not reactive within React
scope. This function has been replaced with another implementation,
that takes React's context and state into account and returns correct
translations. Additionally the function has been made more robust and
to not assume that the input data is valid.