Fix Map stop/Area queries \ Speed them up
Why queries were slow: The map stored the "viewport radius" in meters
in the Redux store, and then the queries asked for stops that were
within a max distance of "radius" from the center point of the map screen.
That might have worked correcly when we were fetching the stops based
on the StopPoint, for witch the locaiton is stores as Geography type
in the db, but on Tiamat side, that locations are stored as Geometry
types, for witch the distance needs to be queried in "degrees" and not
in meters. So in practice the old query always asked for every single
stop within the known universe.

Instead of doing math on multiple ends (calculation "radius",
converting that to "degrees" in UI and calculating distance per stop
on db side) we can just store the actual corner coordinates, aka the
"bounds" of the map, and query for points within that rectangle.