Improve StopRegistry search GroupSelector
Previously the component was implement with HeadlessUI RadioGroup
component. Which work fine when the amount of options is small. But
the implementation ground to halt when searching for stop areas with
'*' as the only search option. Rendering the page took nearly 25s on
my system.

New implementation provides a similar set of features as the HeadlessUI
did, but by foregoing all the generics and compatibility layers of
that, and simply doing the bare minimum manually the new
implementation can render the same dataset in under a second.

Additionally scroll into view features have been added:

Functionality:
* Focuses the selected group on Tab-in
* Scrolls the selected group (back) into view on Tab-out, if the
  element is not expanded.
* Scrolls the selected group (back) into view if the element is
  minimized again.
* Left/Right arrow keys select the previous/next item, and scroll the
  newly focused group into center.
* Up/Down arrow keys do nothing when minimized, but when maximized
  they select the group that is visually above/below the currently
  focused group and scroll the focused item to center.
* Space key selects the currently focused group.