Deps: Add complete @turf as dependency
The amount of needed Turf functions is increasing, and maintaining the
deps list for them is getting tedious. So just add a dependency on the
whole turf bundle instead.

Code changed to import functions from the main bundle,instead of the
sub packages.

Tree shaking should ensure that the endresult app bundle does not grow
and include unused turf functions.