Deps: Bump minor versions & regen yarn.lock
We will "need"(should) to regen the the yarn.lock file when we upgrade
to Eslint V9 in the next commit. But that will cause the package.json
files and the actual yarn.lock to go "out of sync" dues to our use of
`^` in our dependency version declarations.

Next.js tight to version 15.4.6, upgrading it breaks our dev env.