Create extended typed React-router state hook
*Into*: Provides the same typed URl search query state, as previous
`useTypedUrlState` hook does, but additionally provides access to the
state variable of the History API (Router location's `state` field).

*Inspiration*: In previous project I had a hook called:
`useViewState<T>(name, defaultValue)` which provided an API similar to
React's own `useState<T>()`. The hook was used to store simple values
related to the visual state of a given page, where back-and-forth
navigation was expected to happen, and where it was beneficial to
restore the previous visual state after navigation. These were values
such as: Accordion expanded state, active focus element and/or scroll
state of multiple nested scrollable elements. But that implementation
completely bypassed react-router and overrides the whole History API
implementation on the page, to allow both to operate on the same
History stack. So it was somewhat hacky implementation.

*Why the complex unified version*: If a route/page in Jore want's to
store information both on the URL as Search Params, and in the History
state field extra care needs to be taken into account when updating
those, as changing the Search Params will also alter the History State
field, and vice-versa.

For example:
```typescript
// Some wrapper around useLocation().search and useNavigate
const [searchParams, setSearchParams] = useSearchParams(defaultParams);

// Some wrapper around useLocation().state and useNavigate
const [expanded, setExpanded] = useViewState('expanded', false);

const onClickBigRedButton = (newParams) => {
  // Internally this hook would know the location to be:
  // { search: defaultParams, state: { expanded: false } }
  // And change it to:
  // { search: newParams, state: { expanded: false } }
  setSearchParams(newParams);

  // Above call has scheduled a async React state update to happen in
  // the future, but internally this hook still has the old Location
  // stored in memory. Thus, when this updates the expanded value in
  // the Location.state field, it also needs to provide the search
  // field to the navigate() -function. But as the final transition
  // state is calculated synchronously at call time, the previous call
  // to setSearchParams has not yet been flushed in, and this hook
  // still thinks search=defaultParams, and endsup overriding the
  // previous setSearchParams() call.
  setExpanded(true);
}
```

*Features:*
- Same old support for centralized typed and formalized serialization
  and deserialization of URL Search Params.
- Dropped support for extra untyped params. Old `useUrlQuery` and any
  descendant of it must not be used together with the typed state.
- New handles to access and update the Location.state (History.state) field.
- While technically any value that can be cloned with the
  [Structured clone algorithm][1] this implementation only allows
  simple strings, numbers, booleans and compound values (object,
  array) consisting of them.
- No serialization or deserialization.
- Simple shape assertion can and needs to be provided, as technically
  the state could be altered from elsewhere.

[1]: https://developer.mozilla.org/en-US/docs/Web/API/Web_Workers_API/Structured_clone_algorithm