EquipmentReport: Sort by PublicCode + Priority + StartDate
Initial implementation that never ended up as a part of the final PR,
started with a cursor based fetch for the data rows. Thus the data was
ordered by the DB ID "unordered", thus visually unordered in the
generated, but stable and deterministic. It was then up to the user to
sort the rows based on whatever columns they wished in Excel.

But this system does not work reliably for our E2E tests.

While it would be sensible to expect that when making a upsert call:
```
upsert StopPlace {
  id: abc
  quays: [
    { label: A },
    { label: B }
  ]
}
```
The quay labeled __A__ would end up getting DB ID `x` and the one
labeled __B__ would get ID `x+1`. Yet this is not always the case, the
actual order in witch ids `x` and `x+1` get applied (the order in
which the quays are truly inserted into the DB) is instead determined
by which NetexIDs have already been reserved (used). This is due to
Tiamat internally storing the quays in an unordered hash based
collection (HashSet), based on a partial hash of the data. The actual
insertion order ends up being dependent on the Internal state of
Tiamat, that is, what NetexID gets assigned to the Quay.

Therefore, this commit changes the sorting to a more stable order,
that is also visible to the user. Quays are initially sorted and
"grouped" by their public code, then duplicates grouped by priority
and finally, in case of multiple version with same priority by their
start date. This option could end up being less performant due to the
more complex ordering that needs to be performed.