ESlint: Allow for-of-loops
In the base ESLint ruleset that we've inherited (AirBnB) the for-of
loops were forbidden due to "requiring" the use of regenaretor-runtime
for compatibility with IE 11 and older. But native support for the
syntax has existed since 2015 in most major browsers.

Thus, we are not using regenerator-runtime nor transpiling the syntax
and can make free use of it.