Update dependencies
Versions not updated:
--------------------

* Apollo client V3 → V4: A major upgrade
* Headless UI + Tailwind: Major upgrades, separate ticket
* Rollup commonjs/ts plugins: need config changes
* React: Sticking to V8
* Cypress: Major upgrade, needs coordination with node version
* next: Possibly straight upgrade, held for now
* zod: Major upgrade

Other version notes:
--------------------

* react-hook-form: Bound to version 7.66.1, 7.67 has a typing bug,
  which is fixed in main, but has not been published to NPM yet.
* uuid: Bound to version 11.1.0, 12+ versions have problems with
  CJS/MJS import styles, and just don't work properly with our setup.
  Eventually should be replaced with `crypto.randonUUID()` calls, but
  we need HTTPS support in the CI test runs before we can do that.

On Zod:
-------

The react-hooks eslint plugin version has dependency on Zod V4, while
we are still using V3. Normally this would not be a problem, and under
no-packaging-bugs scenario yarn would install the multiple versions of
the package in correct sub-directories and TypeScript and code bundlers
would be able to load the correct version when needed.

But sadly there is a packaging bug in @hookform/resolvers package (at
leas on the old non-updated version we are using), that being a
missing peer-deps declaration to the required zod version. I.e. the
resolver imports zod, but does not declare that it specifically needs
the V3, thus it is possible that under certain scenarios, the V4 of
zod gets installed to the root node_modules, resulting that version
being loaded into the resolves code. Specifically this happens when
trying to build the Docker image.

Fix: Added a dependency to zip v3 to the root package.json of our
repo, this is not a true shared dependency, but this ensures that V3
gets installed as the deduplicated primary version, and v4 gets nested
under the eslint plugin's internal node_modules directory.