UI: Drop direct dep on `@graphql-typed-document-node/core`
Apollo client also has a dependency on it, and the types can and
should be imported trough Apollo instead. Also the only piece of
code that was referencing it was a piece of dead code that has been
dropped completely. One might argue that the code was simply waiting
to be used later on, but the implementation was not valid.
GraphGL `variables` is not the correct way to passtrough the role into
an Apollo query.